Extract executor module
No logical changes.

This will be the foundation of CoreExecutor later.
